translator_agent/pyproject.toml
[project]
name = "translator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "gtts>=2.5.4",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
